Add sign out functionality and update database on every single time user is found.
I understand that this is not a good idea if we think scalability but appwrite at the moment does not allow fallback URL and I have no way to know if the user is coming from a google sign in or magic-url sign in. So at the moment I am updating data everysingle time.