Fixed: Removed all the special symboles from the URL
Removing all the special symbols meang better SEO and also it will save us from the hassel of encoding and decoding the URLs. It also hurts the SEO so that is why there is hypen(-) between all the words and a special 32bit string at the end of the url to uniquely identify the whole URL and article.